Full decoder isolation
