+++ include/aws/mqtt/private/client_impl.h
@@ -10,6 +10,7 @@
 
 #include <aws/mqtt/private/client_impl_shared.h>
 #include <aws/mqtt/private/fixed_header.h>
+#include <aws/mqtt/private/mqtt311_decoder.h>
 #include <aws/mqtt/private/topic_tree.h>
 
 #include <aws/common/hash_table.h>
@@ -117,28 +118,6 @@ enum aws_mqtt_operation_statistic_state_flags {
 
     /* The operation is affecting the connection's "unacked operation" statistics */
     AWS_MQTT_OSS_UNACKED = 1 << 1,
-};
-
-typedef int(
-    packet_handler_fn)(struct aws_byte_cursor message_cursor, void *user_data);
-
-struct aws_mqtt_client_connection_packet_handlers {
-    packet_handler_fn *handlers_by_packet_type[16];
-};
-
-enum aws_mqtt_311_decoder_state_type {
-    AWS_MDST_READ_FIRST_BYTE,
-    AWS_MDST_READ_REMAINING_LENGTH,
-    AWS_MDST_READ_BODY,
-    AWS_MDST_PROTOCOL_ERROR,
-};
-
-struct aws_mqtt311_decoder {
-    const struct aws_mqtt_client_connection_packet_handlers *packet_handlers;
-    enum aws_mqtt_311_decoder_state_type state;
-    size_t total_packet_length;
-    struct aws_byte_buf packet_buffer;
-    void *handler_user_data;
 };
 
 struct aws_mqtt_request {
@@ -437,10 +416,6 @@ void aws_mqtt_connection_statistics_change_operation_statistic_state(
     struct aws_mqtt_request *request,
     enum aws_mqtt_operation_statistic_state_flags new_state_flags);
 
-AWS_MQTT_API void aws_mqtt311_decoder_init(struct aws_mqtt311_decoder *decoder, struct aws_allocator *allocator, void *handler_user_data);
-AWS_MQTT_API void aws_mqtt311_decoder_clean_up(struct aws_mqtt311_decoder *decoder);
-AWS_MQTT_API int aws_mqtt311_decoder_on_bytes_received(struct aws_mqtt311_decoder *decoder, struct aws_byte_cursor data, struct aws_mqtt_client_connection_311_impl *connection);
-AWS_MQTT_API void aws_mqtt311_decoder_reset_for_new_connection(struct aws_mqtt311_decoder *decoder);
-
+AWS_MQTT_API const struct aws_mqtt_client_connection_packet_handlers *aws_mqtt311_get_default_packet_handlers(void);
 
 #endif /* AWS_MQTT_PRIVATE_CLIENT_IMPL_H */
