+++ tests/v3/packet_decoding_tests.c
@@ -5,24 +5,16 @@
 
 #include <aws/testing/aws_test_harness.h>
 
-#include <aws/io/channel_bootstrap.h>
-#include <aws/io/socket.h>
-#include <aws/mqtt/client.h>
-#include <aws/mqtt/private/client_impl.h>
+#include <aws/mqtt/private/fixed_header.h>
+#include <aws/mqtt/private/mqtt311_decoder.h>
+#include <aws/mqtt/private/packets.h>
 
 struct mqtt_311_decoding_test_context {
     struct aws_allocator *allocator;
-    struct aws_client_bootstrap *client_bootstrap;
-    struct aws_event_loop_group *el_group;
-    struct aws_host_resolver *host_resolver;
-    struct aws_mqtt_client *mqtt_client;
-    struct aws_mqtt_client_connection *mqtt_connection;
-    struct aws_socket_options socket_options;
+    struct aws_mqtt311_decoder decoder;
 
     void *expected_packet;
     size_t packet_count;
-
-    struct aws_mqtt_client_connection_packet_handlers test_handlers;
 };
 
 static int s_compare_fixed_header(struct aws_mqtt_fixed_header *expected_header, struct aws_mqtt_fixed_header *actual_header) {
@@ -57,100 +49,93 @@ static int s_decoding_test_handle_publish(struct aws_byte_cursor message_cursor,
     return AWS_OP_SUCCESS;
 }
 
+static struct aws_mqtt_client_connection_packet_handlers s_decoding_test_packet_handlers = {
+    .handlers_by_packet_type = {
+        [AWS_MQTT_PACKET_PUBLISH] = &s_decoding_test_handle_publish,
+    }
+};
+
 static void s_init_decoding_test_context(struct mqtt_311_decoding_test_context *context, struct aws_allocator *allocator) {
     AWS_ZERO_STRUCT(*context);
 
-    aws_mqtt_library_init(allocator);
-
     context->allocator = allocator;
 
-    struct aws_socket_options socket_options = {
-        .connect_timeout_ms = 1000,
-        .domain = AWS_SOCKET_LOCAL,
+    struct aws_mqtt311_decoder_options config = {
+        .packet_handlers = &s_decoding_test_packet_handlers,
+        .handler_user_data = context,
     };
 
-    context->socket_options = socket_options;
-    context->el_group = aws_event_loop_group_new_default(allocator, 1, NULL);
-    struct aws_host_resolver_default_options resolver_options = {
-        .el_group = context->el_group,
-        .max_entries = 1,
-    };
-    context->host_resolver = aws_host_resolver_new_default(allocator, &resolver_options);
-
-    struct aws_client_bootstrap_options bootstrap_options = {
-        .event_loop_group = context->el_group,
-        .user_data = context,
-        .host_resolver = context->host_resolver,
-    };
-
-    context->client_bootstrap = aws_client_bootstrap_new(allocator, &bootstrap_options);
-
-    context->mqtt_client = aws_mqtt_client_new(allocator, context->client_bootstrap);
-    context->mqtt_connection = aws_mqtt_client_connection_new(context->mqtt_client);
-
-    struct aws_mqtt_client_connection_311_impl *connection = context->mqtt_connection->impl;
-    connection->synced_data.state = AWS_MQTT_CLIENT_STATE_CONNECTED;
-
-    context->test_handlers.handlers_by_packet_type[AWS_MQTT_PACKET_PUBLISH] = s_decoding_test_handle_publish;
-
-    connection->thread_data.decoder.packet_handlers = &context->test_handlers;
+    aws_mqtt311_decoder_init(&context->decoder, allocator, &config);
 }
 
 static void s_clean_up_decoding_test_context(struct mqtt_311_decoding_test_context *context) {
-    struct aws_mqtt_client_connection_311_impl *connection = context->mqtt_connection->impl;
-    connection->synced_data.state = AWS_MQTT_CLIENT_STATE_DISCONNECTED;
 
-    aws_mqtt_client_connection_release(context->mqtt_connection);
-    aws_mqtt_client_release(context->mqtt_client);
-
-    aws_client_bootstrap_release(context->client_bootstrap);
-    aws_host_resolver_release(context->host_resolver);
-    aws_event_loop_group_release(context->el_group);
-
-    aws_thread_join_all_managed();
-
-    aws_mqtt_library_clean_up();
+    aws_mqtt311_decoder_clean_up(&context->decoder);
 }
+
+#define TEST_ADJACENT_PACKET_COUNT 4
 
 static int s_mqtt_decode_publish_fn(struct aws_allocator *allocator, void *ctx) {
     (void)ctx;
 
-    struct mqtt_311_decoding_test_context test_context;
-    s_init_decoding_test_context(&test_context, allocator);
+    aws_mqtt_library_init(allocator);
 
-    struct aws_mqtt_client_connection_311_impl *connection = test_context.mqtt_connection->impl;
-    struct aws_mqtt311_decoder *decoder = &connection->thread_data.decoder;
+    /*
+     * For completeness, run the test with payload sizes that lead to a remaining length VLI encoding of 1, 2, 3, and
+     * 4 bytes.
+     */
+    size_t payload_sizes[] = {35, 1234, 1 << 16, 1 << 21};
 
-    struct aws_mqtt_packet_publish publish_packet;
-    ASSERT_SUCCESS(aws_mqtt_packet_publish_init(&publish_packet, true, AWS_MQTT_QOS_AT_LEAST_ONCE, false, aws_byte_cursor_from_c_str("Hello/World"), 12, aws_byte_cursor_from_c_str("Payyyyyyyyyyyyload")));
+    for (size_t i = 0; i < AWS_ARRAY_SIZE(payload_sizes); ++i) {
+        struct mqtt_311_decoding_test_context test_context;
+        s_init_decoding_test_context(&test_context, allocator);
 
-    test_context.expected_packet = &publish_packet;
-    
-    struct aws_byte_buf encoded_buffer;
-    aws_byte_buf_init(&encoded_buffer, allocator, 16384);
+        struct aws_mqtt311_decoder *decoder = &test_context.decoder;
 
-    ASSERT_SUCCESS(aws_mqtt_packet_publish_encode(&encoded_buffer, &publish_packet));
-    ASSERT_SUCCESS(aws_mqtt_packet_publish_encode(&encoded_buffer, &publish_packet));
-    ASSERT_SUCCESS(aws_mqtt_packet_publish_encode(&encoded_buffer, &publish_packet));
-    ASSERT_SUCCESS(aws_mqtt_packet_publish_encode(&encoded_buffer, &publish_packet));
+        size_t publish_payload_size = payload_sizes[i];
+        AWS_VARIABLE_LENGTH_ARRAY(uint8_t, payload_buffer, publish_payload_size);
 
-    size_t fragment_lengths[] = { 1, 2, 3, 5, 7, 11, 23, 37, 67, 128};
+        struct aws_mqtt_packet_publish publish_packet;
+        ASSERT_SUCCESS(aws_mqtt_packet_publish_init(
+            &publish_packet,
+            true,
+            AWS_MQTT_QOS_AT_LEAST_ONCE,
+            false,
+            aws_byte_cursor_from_c_str("Hello/World"),
+            12,
+            aws_byte_cursor_from_array(payload_buffer, publish_payload_size)));
 
-    for (size_t i = 0; i < AWS_ARRAY_SIZE(fragment_lengths); ++i) {
-        size_t fragment_length = fragment_lengths[i];
+        test_context.expected_packet = &publish_packet;
 
-        struct aws_byte_cursor packet_cursor = aws_byte_cursor_from_buf(&encoded_buffer);
-        while (packet_cursor.len > 0) {
-            size_t advance = aws_min_size(packet_cursor.len, fragment_length);
-            struct aws_byte_cursor fragment_cursor = aws_byte_cursor_advance(&packet_cursor, advance);
+        struct aws_byte_buf encoded_buffer;
+        aws_byte_buf_init(&encoded_buffer, allocator, (publish_payload_size + 100) * TEST_ADJACENT_PACKET_COUNT );
 
-            ASSERT_SUCCESS(aws_mqtt311_decoder_on_bytes_received(decoder, fragment_cursor, connection));
+        for (size_t j = 0; j < TEST_ADJACENT_PACKET_COUNT; ++j) {
+            ASSERT_SUCCESS(aws_mqtt_packet_publish_encode(&encoded_buffer, &publish_packet));
         }
+
+        size_t fragment_lengths[] = {1, 2, 3, 5, 7, 11, 23, 37, 67, 131};
+
+        for (size_t j = 0; j < AWS_ARRAY_SIZE(fragment_lengths); ++j) {
+            size_t fragment_length = fragment_lengths[j];
+
+            struct aws_byte_cursor packet_cursor = aws_byte_cursor_from_buf(&encoded_buffer);
+            while (packet_cursor.len > 0) {
+                size_t advance = aws_min_size(packet_cursor.len, fragment_length);
+                struct aws_byte_cursor fragment_cursor = aws_byte_cursor_advance(&packet_cursor, advance);
+
+                ASSERT_SUCCESS(aws_mqtt311_decoder_on_bytes_received(decoder, fragment_cursor));
+            }
+        }
+
+        ASSERT_INT_EQUALS(4 * AWS_ARRAY_SIZE(fragment_lengths), test_context.packet_count);
+
+        aws_byte_buf_clean_up(&encoded_buffer);
+
+        s_clean_up_decoding_test_context(&test_context);
     }
 
-    ASSERT_INT_EQUALS(4 * AWS_ARRAY_SIZE(fragment_lengths), test_context.packet_count);
-
-    s_clean_up_decoding_test_context(&test_context);
+    aws_mqtt_library_clean_up();
 
     return AWS_OP_SUCCESS;
 }
